requirements.txt
requests==2.25.1
requests-oauthlib==1.3.1
panda==0.3.1
pandas==1.2.2
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
discord==1.7.3
discord.py==1.7.3
python-dotenv~=1.0.1